Analysis featuring different configurations of model
How does changing the activation function affect the accuracy? What if the epochs are increased to a huge number? Visualize how the model trains differently making different configurations in the model. See conditions where the model overfits/ underfits. Find out the parameters of loss function, activation function, no. of hidden layers, type of hidden layers, dropout, no. of epochs for the best fit of the model.